fixed reversed geocoding and getCurrentPosition()
Had to get an api key for the geocoding and set the page to run client side to get the position of the user. Also changed api to Geoapify so that the api calls can occur in the browser. I also added a change to the dev script so development testing now happens in https. Had to get some dummy certs for it (provided by Next,js).